Structure for interactive functionalities
A new view is dedicated to interactive fonctionalities. The view is called inputline.
A structure input is dedicated to interactive handlers. The structure contains a channel and
the content of the buffer.
The channel is used to stop the functionality handler after a post-traitement and a call
to currTopViewHandler on inputline. Then, when pressing the Enter key, the validateInput function
will close the input view and place the buffer content into the structure input. At the end,
validateInput will resume the functionality handler function in order to do the specific stuffs